Convert booking service to payment initiator
