+++ HotelBackend.sln
@@ -9,7 +9,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DataAccess", "DataAccess\Da
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HotelBackend.ReservationService", "HotelBackend.ReservationService\HotelBackend.ReservationService.csproj", "{67B1BE67-6D77-4B19-9935-7CEE9D7E1A9D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HotelBackend.BookingService", "HotelBackend.BookingService\HotelBackend.BookingService.csproj", "{5FD13D4F-CE21-4FA4-9B2E-ED83DA8BBBD1}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HotelBackend.PaymentInitiator", "HotelBackend.PaymentInitiator\HotelBackend.PaymentInitiator.csproj", "{5FD13D4F-CE21-4FA4-9B2E-ED83DA8BBBD1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
